LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/navigation_unload_same_origin-expected.txt

frame "<!--framePath //<!--frame0-->-->" - has 1 onunload handler(s)

FAIL Same-origin navigation started from unload handler assert_equals: expected "001-3" but got "001-2"
